Pass next value to unhook().
The patch adds a third parameter to unhook() - the next value of
the property, or undefined if the hook is removed.

This is consistent with hook(), which receives the previous value
as the third parameter.

As of 4aedb4a3697b7e3f3eee11f769e20a02de5c534f, unhook() is invoked
every time the hook gets patched. The next value can be used to determine
whether unhook() can avoid doing uneccessary work, for example:

    MyHook.prototype.unhook = function(node, prop, next) {
      if (next && next.value === this.value) {
        return
      }
      // ... do work ...
    }