Remove redundant code branch in diffProps.
I guess that the original idea was to put a 'continue' statement
to the bottom of that if branch.

While that would be more performant in the case where the aKey property
is missing from b, it is slower in the more common case where the aKey
property is still present in b, so I think it makes sense to just remove
the if branch altogether.